+++ subvt-persistence/src/postgres/network/extrinsic/payout_stakers.rs
@@ -54,7 +54,7 @@ impl PostgreSQLNetworkStorage {
             .bind(maybe_nesting_index)
             .bind(caller_account_id.to_string())
             .bind(validator_account_id.to_string())
-            .bind(era_index)
+            .bind(era_index as i64)
             .bind(is_successful)
             .fetch_optional(&self.connection_pool)
             .await?;
